refactor: extract RecipeCard into it's own component
